requirements.txt
PyGithub
bs4
requests